Server side render 404 to prevent hydration errors
